// src/app/page/master/metode-perhitungan/metode-perhitungan.interface.ts
import {EventEmitter} from '@angular/core';

export interface MetodePerhitungan {
	kode?: any;
	id?: any;
	mataUang?: any;
	kdProfile?: string;
	namaModulAplikasi?: string;
	version?: string;
	reportDisplay?: string;
	kodeExternal?: string;
	namaExternal?: string;
	statusEnabled?: boolean;
	noRec?: string;
	label?: any;
	value?: any;
}